intValue化するとどうなるかテスト

+++ myLib.MIT/myLib.container.mkv/src/test/java/com/ttProject/container/mkv/test/ListTest.java
@@ -26,7 +26,7 @@ public class ListTest {
 		list.add(5);
 		list.add(2);
 		logger.info(list);
-		logger.info(list.remove((Integer)1));
+		logger.info(list.remove(((Integer)1).intValue()));
 		logger.info(list);
 	}
 }
